Bump anomaly-detection version to 1.2
