+++ src/test/java/org/opensearch/ad/transport/ADStatsITTests.java
@@ -13,16 +13,14 @@ package org.opensearch.ad.transport;
 
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.concurrent.ExecutionException;
 
 import org.opensearch.ad.AnomalyDetectorPlugin;
-import org.opensearch.ad.model.DetectorProfileName;
 import org.opensearch.plugins.Plugin;
 import org.opensearch.test.OpenSearchIntegTestCase;
 
 @OpenSearchIntegTestCase.ClusterScope(transportClientRatio = 0.9)
-public class ProfileIT extends OpenSearchIntegTestCase {
+public class ADStatsITTests extends OpenSearchIntegTestCase {
 
     @Override
     protected Collection<Class<? extends Plugin>> nodePlugins() {
@@ -34,10 +32,10 @@ public class ProfileIT extends OpenSearchIntegTestCase {
         return Collections.singletonList(AnomalyDetectorPlugin.class);
     }
 
-    public void testNormalProfile() throws ExecutionException, InterruptedException {
-        ProfileRequest profileRequest = new ProfileRequest("123", new HashSet<DetectorProfileName>(), false);
+    public void testNormalADStats() throws ExecutionException, InterruptedException {
+        ADStatsRequest adStatsRequest = new ADStatsRequest(new String[0]);
 
-        ProfileResponse response = client().execute(ProfileAction.INSTANCE, profileRequest).get();
-        assertTrue("getting profile failed", !response.hasFailures());
+        ADStatsNodesResponse response = client().execute(ADStatsNodesAction.INSTANCE, adStatsRequest).get();
+        assertTrue("getting stats failed", !response.hasFailures());
     }
 }
